requirements.txt
pandas==1.5.3
ptflops==0.6.9
PyYAML==6.0
timm==0.6.12
torch==1.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tqdm==4.64.1
